feat: splits projects and adds solution file
before the project wasn't a library, so not particularily very useful

this change splits what was there into the library component, and the
example component

all of this referenced in a solution file to make the build process a
bit cleaner

the interface is fairly nice, the example has basically zero
dependencies, apart from the other project.

Now if someone wants to use the library, they can build the solution
which will build both the library and the example, if they don't want to
verify the tests, these can be ignored and the library dll can be used
in the project.

I had to nest everything in folders to make the compiler happy, dumping
everything in one big folder seems to confuse it for some reason. I
should have done a bit more organising on disk any way