removes Invalid FunctionCode, adds more deserialization
The Invalid FunctionCode and Exception code is outside of the modbus
specification.

It was a code-smell to have it in

A much better way is to remain in-spec with the standard, but use a
result type to capture the garbage data

Consequently, the map request to response function now responds to
un-parsable data by killing the connection with the client.

Furthermore, this forced improvements with the validatePdu function,
yielding Modbus Error types that reflect the request type (inclusion of
the ExceptionCode is still not covered, however this brings us a lot
closer)

I will probably update the TryParse methods, to return Result<Request,
ExceptionCode * exn> instead of Result<Request, PDU * exn>,
This way knowledge of what's valid for a function code stays with the
type.

I will probably change mod funcs to the type Result<_Response,
ExceptionCode>, so that ModErrors can be raised by the business logic,
(i.e. address out of range)