broken: code duplication in request and response types reduced
(it was working before i started changing the mbap interface - should
have committed sooner. Cherrypicking was too much effort)

I'm committing broken and *early* because i want to continue working on
this code on my other device.

the use of "partialSerialization" is used for the repeated types, this
is because the repeated type (i.e. ReqOffQuant) doesn't have a
one-to-one mapping to a particular functioncode, thus this information
needs to be provided elsewhere. Since the interface within the code is
request or response, which is the DU of the various function codes, when
calling serialize on the Response or Request type, the request or
response type 'knows' whether it is partial or not and can append the
correct function code

I decided that this is a ModbusTCP library (although large chunks of it
'can' be recycled for serial modbus)

I've decided that the delegate that 'does all the things' should work
with the MBap type, and return an MBap type. In this commit, there's one
MBap type for both request and response. This is *wrong* and needs to be
changed. Instead, there will be a *request mbap* and *response mbap*
type

parsing into the 'master' mbap type means that we are not 'throwing-away'
any information, so we can have multiple modbus slaves, (which would be
handy if we wanted to build a modbusTCP to modbusRTU bridge)

Furthermore, there is essentially one 'entry-point' whereby we
deserialize and expect to get an MBAP type, what's nice is we can call
'parse' or 'serialize' at the top type, and all this has to do is call
it on each 'child' type