feat: The library is MBap Request and Response focused
The library is unopinionated about how to work with modbus

The user should supply a delegate of MBapRequest to Result<MBapResponse,
()>

This can probably be more useful to use exception, etc, but there's
enough here to get started.

What i don't like about this library is i can't find a way to keep the
memory usage down, on my laptop (ubuntu with 16GB ram) the debug version
of the program will use up to 1.7GB. I know this business-as-usual with
dotnet but i can see no reason why the ram usage is so high!

I was smashing the server with up to 400 requests per second (reading
and writing) and CPU usage was well below 10%.

I would like to compile this for and run it on a raspberry pi, i think
this could work very nicely with an rPi

The next step is to write a basic client, since all of the serialization
and deserialization has been taken care of this should be fairly simple.

I will try to make smaller, better documented commits

ToDo:
Write the client
Restore testing
Improve 'exception handling'
Add logging
Add round-trip testing