improv: handleReceive scoped to build func
The handleReceive function is only ever called by the handle dispatcher
loop (one thread spawns a new thread every time there is a new modbus
connection), instead of setting the function to private, it has been
moved inside the build scope.

The server function was put inside a 'Server' module. The server
function was renamed 'build' since the function, 'builds' the server.

Furthermore, as i intend to add a client to this library, it makes sense
to put the client and the server in separate modules, depending on the
size of the client, these may be split into two files