imprv: the types better reflect the intent
before the types were modeled on the modbus standard instead of the
intent of their use.

i.e. a response to a request for 6 coils, shouldn't have 8 bool values +
a quantity of 6. This is maddness! the serializer and deserializer can
take care of this for us.

Consequently, we're not carrying around additional unncessary bits in
our model.