feat(client): the client returns the expected quantity
The client was returning the deserialized result from the server.

Thus, the result would consist of padding to 1 byte's worth of bits.

A small bit of smarts was added to the client, so that it checks how
many bits were requested in the first place and returns that.

The client can't rely on the return payload, because the return payload
doesn't include the number of bits requested