improves readme
the readme did not include info about the docker image.

since this is now being published to docker hub, the readme needs to
include information pertaining to both the source and the docker image.

gives very simple instructions to get started with a modbus server, or
how to get started, using the example code.